+++ src/components/PrivatePage/PrivatePage.js
@@ -1,32 +1,56 @@
 
-import React, { Component, PropTypes } from 'react';
+import React, { Component, PropTypes } from 'react'
 
-import withStyles from 'isomorphic-style-loader/lib/withStyles';
-import s from './PrivatePage.scss';
+import withStyles from 'isomorphic-style-loader/lib/withStyles'
+import s from './PrivatePage.scss'
 
-import withAuth from '../withAuth'
+import TodoStore from '../../stores/TodoStore'
+import withAuth from '../../decorators/withAuth'
 
+@withAuth
 class PrivatePage extends Component {
 
   static contextTypes = {
     onSetTitle: PropTypes.func.isRequired,
   };
 
+  constructor(props) {
+    super(props)
+    this.state = TodoStore.getState()
+  };
+
   componentWillMount() {
-    this.context.onSetTitle('Hello ' + this.props.userProfile.personname);
-  }
+    this.context.onSetTitle('Hello ' + this.props.userProfile.name)
+  };
+
+  componentDidMount() {
+    TodoStore.listen(this.onChange)
+  };
+
+  componentWillUnmount() {
+    TodoStore.unlisten(this.onChange)
+  };
+
+  onChange = (state) => {
+    this.setState(state)
+  };
 
   render() {
     const profile = this.props.userProfile
 
     return (
       <div className={s.root}>
-        <h1>This is example of prtected page.</h1>
+        <h1>This is example of prtected page</h1>
         <p>You profile info:</p>
         <table>
+        <tbody>
           <tr>
-            <td>Provider</td>
-            <td>{profile.service}</td>
+            <td>Login via</td>
+            <td>{profile.provider}</td>
+          </tr>
+          <tr>
+            <td>User id</td>
+            <td>{profile.id}</td>
           </tr>
           <tr>
             <td>Avatar</td>
@@ -34,17 +58,48 @@ class PrivatePage extends Component {
           </tr>
           <tr>
             <td>Person</td>
-            <td>{profile.personname}</td>
+            <td>{profile.name}</td>
           </tr>
           <tr>
-            <td>Access token:</td>
+            <td>Oauth token</td>
             <td>{profile.token}</td>
           </tr>
+        </tbody>
         </table>
+
+        <h1>Todo list</h1>
+        {this.renderTodoList()}
+
       </div>
     );
   }
 
+  renderTodoList() {
+    const st = this.state
+
+    if (st.loading) {
+      return <span>loading...</span>
+    }
+
+    if (st.errorMessage) {
+      return <span>{st.errorMessage}</span>
+    }
+
+    if (st.todos.length == 0) {
+      return <span>list is empty</span>
+    }
+
+    return (
+      <ul>
+        {
+          st.todos.map((item, idx) => {
+            return <li key={idx}>{item}</li>
+          })
+        }
+      </ul>
+    )
+  }
+
 }
 
-export default withStyles(withAuth(PrivatePage), s);
+export default withStyles(PrivatePage, s);