+++ CHANGELOG.md
@@ -1,1 +1,11 @@
 ## Change Log
+
+Init auth-starter-kit:
+
+* server-client redirects
+* TodoStore example
+* Google/facebook/twitter auth (see passportjs)
+* css @media queries support
+* component generator tool (see doc.)
+* component decorators support
+* Client-side cookie util